packages/qt-shared/pyproject.toml
[project]
name = "qt-shared"
version = "0.1.0"
description = "Shared infrastructure for QueryTorque - auth, billing, database, LLM clients"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "QueryTorque", email = "support@querytorque.com"}
]
keywords = ["querytorque", "shared", "infrastructure"]

dependencies = [
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Billing
    "stripe>=8.0.0",
    # LLM clients (optional - installed as extras)
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.39.0"]
groq = ["groq>=0.11.0"]
openai = ["openai>=1.50.0"]
gemini = ["google-generativeai>=0.8.0"]
all-llm = [
    "anthropic>=0.39.0",
    "groq>=0.11.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["qt_shared"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
